feat: complete instance scoping, lifecycle hardening, runtime diagnostics and asset loading (#1177)
* feat: complete instance scoping for metadata mutations, message queries, producers and dashboard

Consolidates #1112, #1120, #1162, #1166, #1177: route Apache Topic mutations,
test messages, producer connections, message queries, dashboard data and
metrics sources by the selected instance; scope metadata deletes to the
current cluster; redirect unknown instance routes.

* fix: harden instance lifecycle, runtime admin caches and cloud credential invalidation

Consolidates #1185, #1213, #1225, #1236, #1238: release unused runtime
AdminClients, normalize NameServer cache keys, align cache lifecycle,
validate Apache instance endpoints and type before persistence, invalidate
Aliyun clients after credential changes, expose unavailable cloud resource
counts and check current resources before deleting an instance.

* fix: harden runtime diagnostics, dashboard overview and async detail freshness

Consolidates #1089, #1205, #1219, #1223, #1240: guard stale async detail
responses (consumer groups, LiteTopic), coordinate group refreshes, remove
unsupported group/subscription actions, surface unavailable broker runtime
statistics/versions and cluster discovery failures, wire Broker Cluster
restart, mark dashboard clusters warning on runtime failures, preserve TPS
precision, reuse broker runtime snapshots, distinguish unavailable DLQ
statistics, handle missing DLQ enqueue timestamps, stop stalled pull scans
and report DLQ resend failures.

* fix: harden bundled asset loading, metrics labels and defaults; docs quick start; align tests

Consolidates #1122, #1123, #1200, #1217: decode Grafana dashboard assets as
UTF-8, skip invalid Grafana assets and incomplete alert rules, bound
Prometheus time-series labels, stop seeding demo resources by default and
complete the Docker Compose quick start docs. Also aligns pre-existing tests
with the integrated behaviors (required instance type, live resource counts
on delete, long TPS fields, default dashboard cluster type).

+++ server/src/main/java/org/apache/rocketmq/studio/ops/dashboard/ClusterOverviewVO.java
@@ -39,8 +39,8 @@ public class ClusterOverviewVO {
     private int proxies;
     private int topics;
     private int groups;
-    private int tpsIn;
-    private int tpsOut;
+    private long tpsIn;
+    private long tpsOut;
     private String version;
     private List<Integer> throughput;
 }
